💄[style] axios 위치 변경
