+++ src/components/TourList/TourListItemSection/TourListItemSection.jsx
@@ -7,24 +7,11 @@ import Icon from '../../layout/atom/Icon';
 import TourListCardSquare from '../TourListCardSquare';
 
 function TourListItemSection() {
-  const [tourList, setTourList] = useState();
-
-  useEffect(() => {
-    getTourList();
-  }, []);
-
-  async function getTourList() {
-    try {
-      const response = await axios.get(
-        `http://15.165.135.183:8080/api/tour/filter?city=%ED%8C%8C%EB%A6%AC&order=%EC%B6%94%EC%B2%9C%EC%88%9C&minimumPrice=100&maximumPrice=100000000000&tourType=all&page=1`,
-      );
-      const data = response.data.data;
-      setTourList(data.tourList);
-      console.log(tourList);
-    } catch (error) {
-      console.error(error);
-    }
-  }
+  const [sort, setSort] = useState();
+  const [minimumPrice, setminiMumPrice] = useState();
+  const [maximumPrice, setMaximumPrice] = useState();
+  const [tourType, setTourType] = useState();
+  const [page, setPage] = useState();
   return (
     <>
       <Flex
